+++ Modules/Intent.Modules.CSharp.Importer/Tasks/FileDirectoryPreviewTask.cs
@@ -1,0 +1,8 @@
+﻿using Intent.Modules.Shared.Importer.Tasks;
+
+namespace Intent.Modules.CSharp.Importer.Tasks;
+
+public class FileDirectoryPreviewTask : FileDirectoryPreviewTaskBase
+{
+    public override string TaskTypeId => "Intent.Modules.Importer.CSharp.FileDirectoryPreviewTask";
+}