+++ Modules/Intent.Modules.Shared.FileImporter/Task/Models/ExecuteResult.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 
-namespace Intent.Modules.Json.Importer.Tasks.Models;
+namespace Intent.Modules.Shared.Importer.Tasks.Models;
 
 public class ExecuteResult<T>(T? result)
 {