+++ Modules/Intent.Modules.Shared.FileImporter/Task/Models/Models.cs
@@ -1,4 +1,4 @@
-namespace Intent.Modules.Json.Importer.Tasks.Models;
+namespace Intent.Modules.Shared.Importer.Tasks.Models;
 
 public class FilePreviewInputModel
 {