+++ Modules/Intent.Modules.OpenApi.Importer/release-notes.md
@@ -1,3 +1,7 @@
+### Version 1.1.13
+
+- Fixed: Internal refactor to ensure that `Intent.Modules.Importer.FileDirectoryPreviewTask` no longer exists as other importers who have shared code could end up registering the same Task ID and cause a crash.
+
 ### Version 1.1.12
 
 - Improvement: Added ProjectUrl link.